Refactor MeshFederation controller
