+++ internal/controller/meshfederation/meshfederation_controller.go
@@ -16,26 +16,17 @@ package meshfederation
 
 import (
 	"context"
-	"fmt"
 
-	networkingspecv1alpha3 "istio.io/api/networking/v1alpha3"
-	securityspecv1beta1 "istio.io/api/security/v1beta1"
-	typev1beta1 "istio.io/api/type/v1beta1"
 	networkingv1alpha3 "istio.io/client-go/pkg/apis/networking/v1alpha3"
 	securityv1beta1 "istio.io/client-go/pkg/apis/security/v1beta1"
 	corev1 "k8s.io/api/core/v1"
 	"k8s.io/apimachinery/pkg/api/errors"
-	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
-	"k8s.io/apimachinery/pkg/labels"
 	"k8s.io/apimachinery/pkg/types"
 	ctrl "sigs.k8s.io/controller-runtime"
 	"sigs.k8s.io/controller-runtime/pkg/builder"
 	"sigs.k8s.io/controller-runtime/pkg/client"
-	"sigs.k8s.io/controller-runtime/pkg/controller/controllerutil"
-	"sigs.k8s.io/controller-runtime/pkg/event"
 	"sigs.k8s.io/controller-runtime/pkg/handler"
 	"sigs.k8s.io/controller-runtime/pkg/log"
-	"sigs.k8s.io/controller-runtime/pkg/predicate"
 	"sigs.k8s.io/controller-runtime/pkg/reconcile"
 
 	"github.com/openshift-service-mesh/federation/api/v1alpha1"
@@ -99,27 +90,7 @@ func (r *Reconciler) Reconcile(ctx context.Context, req ctrl.Request) (ctrl.Resu
 
 	r.instance = instance
 
-	// ensure strict mTLS for local FDS server
-	// TODO: do not update if not necessary
-	peerAuth := &securityv1beta1.PeerAuthentication{
-		ObjectMeta: metav1.ObjectMeta{
-			Name:      "fds-strict-mtls",
-			Namespace: r.namespace,
-		},
-	}
-	if _, err := controllerutil.CreateOrUpdate(ctx, r.Client, peerAuth, func() error {
-		peerAuth.Spec = securityspecv1beta1.PeerAuthentication{
-			Selector: &typev1beta1.WorkloadSelector{
-				MatchLabels: map[string]string{
-					"app.kubernetes.io/name": "federation-controller",
-				},
-			},
-			Mtls: &securityspecv1beta1.PeerAuthentication_MutualTLS{
-				Mode: securityspecv1beta1.PeerAuthentication_MutualTLS_STRICT,
-			},
-		}
-		return controllerutil.SetControllerReference(instance, peerAuth, r.Scheme())
-	}); err != nil {
+	if err := r.reconcilePeerAuthentication(ctx); err != nil {
 		logger.Error(err, "failed to create or update peer authentication")
 		return ctrl.Result{}, err
 	}
@@ -138,43 +109,13 @@ func (r *Reconciler) Reconcile(ctx context.Context, req ctrl.Request) (ctrl.Resu
 		}()
 	}
 
-	// Send federated services
-	exportedServices := &corev1.ServiceList{}
-	// TODO: Add support for matchExpressions
-	if err := r.Client.List(context.Background(), exportedServices, client.MatchingLabels(r.instance.Spec.ExportRules.ServiceSelectors.MatchLabels)); err != nil {
-		return ctrl.Result{}, fmt.Errorf("failed to list services: %w", err)
+	federatedServices, svcErr := r.reconcileFederatedServices(ctx)
+	if svcErr != nil {
+		logger.Error(svcErr, "failed to reconcile federated services")
+		return ctrl.Result{}, svcErr
 	}
-	r.pushRequests <- xds.PushRequest{TypeUrl: xds.ExportedServiceTypeUrl}
 
-	// Expose federated services
-	// TODO: do not update if not necessary
-	hosts := []string{fmt.Sprintf("federation-discovery-service-%s.%s.svc.cluster.local", instance.Name, instance.Namespace)}
-	for _, svc := range exportedServices.Items {
-		hosts = append(hosts, fmt.Sprintf("%s.%s.svc.cluster.local", svc.Name, svc.Namespace))
-	}
-	gateway := &networkingv1alpha3.Gateway{
-		ObjectMeta: metav1.ObjectMeta{
-			Name:      "federation-ingress-gateway",
-			Namespace: r.namespace,
-		},
-	}
-	if _, err := controllerutil.CreateOrUpdate(ctx, r.Client, gateway, func() error {
-		gateway.Spec = networkingspecv1alpha3.Gateway{
-			Selector: r.instance.Spec.IngressConfig.GatewayConfig.Selector,
-			Servers: []*networkingspecv1alpha3.Server{{
-				Hosts: hosts,
-				Port: &networkingspecv1alpha3.Port{
-					Number:   r.instance.Spec.IngressConfig.GatewayConfig.PortConfig.Number,
-					Name:     r.instance.Spec.IngressConfig.GatewayConfig.PortConfig.Name,
-					Protocol: "TLS",
-				},
-				Tls: &networkingspecv1alpha3.ServerTLSSettings{
-					Mode: networkingspecv1alpha3.ServerTLSSettings_AUTO_PASSTHROUGH,
-				},
-			}},
-		}
-		return controllerutil.SetControllerReference(instance, peerAuth, r.Scheme())
-	}); err != nil {
+	if err := r.reconcileGateway(ctx, federatedServices); err != nil {
 		logger.Error(err, "failed to create or update peer authentication")
 		return ctrl.Result{}, err
 	}
@@ -186,53 +127,22 @@ func (r *Reconciler) Reconcile(ctx context.Context, req ctrl.Request) (ctrl.Resu
 func (r *Reconciler) SetupWithManager(mgr ctrl.Manager) error {
 	return ctrl.NewControllerManagedBy(mgr).
 		For(&v1alpha1.MeshFederation{}).
+		Owns(&securityv1beta1.PeerAuthentication{}).
+		Owns(&networkingv1alpha3.Gateway{}).
 		Watches(
 			&corev1.Service{},
-			// TODO: this function will not work properly when more than 1 MeshFederation resource exists
-			handler.EnqueueRequestsFromMapFunc(func(ctx context.Context, object client.Object) []reconcile.Request {
-				instances := &v1alpha1.MeshFederationList{}
-				// TODO: How can we handle this error?
-				if err := r.Client.List(ctx, instances); err != nil {
-					return []reconcile.Request{}
-				}
-				return []reconcile.Request{{NamespacedName: types.NamespacedName{
-					Name:      instances.Items[0].Name,
-					Namespace: instances.Items[0].Namespace,
-				}}}
-			}),
-			builder.WithPredicates(
-				predicate.Funcs{
-					CreateFunc: func(e event.CreateEvent) bool {
-						return r.matchesExportRules(e.Object.(*corev1.Service))
-					},
-					UpdateFunc: func(e event.UpdateEvent) bool {
-						oldSvc := e.ObjectOld.(*corev1.Service)
-						newSvc := e.ObjectNew.(*corev1.Service)
-						return r.matchesExportRules(oldSvc) != r.matchesExportRules(newSvc)
-					},
-					DeleteFunc: func(e event.DeleteEvent) bool {
-						return r.matchesExportRules(e.Object.(*corev1.Service))
-					},
-					GenericFunc: func(e event.GenericEvent) bool {
-						return false
-					},
-				},
-			),
+			handler.EnqueueRequestsFromMapFunc(r.enqueueRequestForCurrentInstance),
+			builder.WithPredicates(r.enqueueIfMatchExportRules()),
 		).
 		Complete(r)
 }
 
-func (r *Reconciler) matchesExportRules(svc *corev1.Service) bool {
+func (r *Reconciler) enqueueRequestForCurrentInstance(_ context.Context, _ client.Object) []reconcile.Request {
 	if r.instance == nil {
-		return false
+		return []reconcile.Request{}
 	}
-	if r.instance.Spec.ExportRules == nil {
-		return false
-	}
-	if r.instance.Spec.ExportRules.ServiceSelectors == nil {
-		return true
-	}
-	// TODO: add support for matchExpressions
-	selector := labels.SelectorFromSet(r.instance.Spec.ExportRules.ServiceSelectors.MatchLabels)
-	return selector.Matches(labels.Set(svc.GetLabels()))
+	return []reconcile.Request{{NamespacedName: types.NamespacedName{
+		Name:      r.instance.Name,
+		Namespace: r.instance.Namespace,
+	}}}
 }
